+++ core/js/typespecs/function_id.ts
