+++ core/js/typespecs/index.ts
@@ -3,6 +3,6 @@ export * from "./app_types";
 export * from "./common_types";
 export * from "./custom_types";
 export * from "./functions";
-export * from "./function-id";
+export * from "./function_id";
 export * from "./prompt";
 export * from "./view";
